refactor functions_loader and interface module.
